admin verification and file upload

+++ routes/promoRouter.js
@@ -21,7 +21,8 @@ promoRouter.route("/")
     .catch(err =>  next(err))
 })
 //POST creates a new resource
-.post(authenticate.verifyUser,(req, res, next)=>{
+.post(authenticate.verifyUser,authenticate.verifyAdmin,(req, res, next)=>{
+    authenticate.verifyAdmin(req, res, next);
     Promotions.create(req.body)
     .then(promotions =>{
         res.statusCode = 200;
@@ -30,11 +31,12 @@ promoRouter.route("/")
     }, (err) => next(err))
     .catch( (err) => next(err));
 })
-.put(authenticate.verifyUser,(req, res, next)=>{
+.put(authenticate.verifyUser,authenticate.verifyAdmin,(req, res, next)=>{
     res.statusCode = 403;
     res.end("PUT operation not supported on /promos");
 })
-.delete(authenticate.verifyUser,(req, res, next)=>{
+.delete(authenticate.verifyUser,authenticate.verifyAdmin,(req, res, next)=>{
+    authenticate.verifyAdmin(req, res, next);
     Promotions.remove({})
     .then(resp =>{
         res.status = 200;
@@ -56,12 +58,13 @@ promoRouter.route("/:promoId")
     }, (err) => next(err))
     .catch(err => next(err));
 })
-.post(authenticate.verifyUser,(req, res, next)=>{
+.post(authenticate.verifyUser,authenticate.verifyAdmin,(req, res, next)=>{
     res.statusCode = 403;
     res.end("POST operation not supported on /promotions/"
         + req.params.promoId);
 })
-.put(authenticate.verifyUser,(req, res, next)=>{
+.put(authenticate.verifyUser,authenticate.verifyAdmin,(req, res, next)=>{
+    authenticate.verifyAdmin(req, res, next);
     Promotions.findByIdAndUpdate(req.params.promoId, {$set: req.body}, {new: true})
     .then(promotion =>{
         console.log("Promotion Updated", promotion._doc);
@@ -71,7 +74,8 @@ promoRouter.route("/:promoId")
     }, (err) => next(err))
     .catch(err => next(err));
 })
-.delete(authenticate.verifyUser,(req, res, next)=>{
+.delete(authenticate.verifyUser,authenticate.verifyAdmin,(req, res, next)=>{
+    authenticate.verifyAdmin(req, res, next);
     Promotions.findByIdAndRemove(req.params.promoId)
     .then(promotion =>{
         console.log("Promotion Removed!", promotion._doc)
